fix: better sub object handling
