--- internal/project/repository/eventsourcing/model_test.go
@@ -1,1114 +1,0 @@
-package eventsourcing
-
-import (
-	"encoding/json"
-	"testing"
-
-	es_models "github.com/caos/zitadel/internal/eventstore/models"
-	"github.com/caos/zitadel/internal/project/model"
-)
-
-func TestProjectChanges(t *testing.T) {
-	type args struct {
-		existing *Project
-		new      *Project
-	}
-	type res struct {
-		changesLen int
-	}
-	tests := []struct {
-		name string
-		args args
-		res  res
-	}{
-		{
-			name: "project name changes",
-			args: args{
-				existing: &Project{Name: "Name"},
-				new:      &Project{Name: "NameChanged"},
-			},
-			res: res{
-				changesLen: 1,
-			},
-		},
-		{
-			name: "no changes",
-			args: args{
-				existing: &Project{Name: "Name"},
-				new:      &Project{Name: "Name"},
-			},
-			res: res{
-				changesLen: 0,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			changes := tt.args.existing.Changes(tt.args.new)
-			if len(changes) != tt.res.changesLen {
-				t.Errorf("got wrong changes len: expected: %v, actual: %v ", tt.res.changesLen, len(changes))
-			}
-		})
-	}
-}
-
-func TestApplicationChanges(t *testing.T) {
-	type args struct {
-		existing *Application
-		new      *Application
-	}
-	type res struct {
-		changesLen int
-	}
-	tests := []struct {
-		name string
-		args args
-		res  res
-	}{
-		{
-			name: "application name changes",
-			args: args{
-				existing: &Application{AppID: "AppID", Name: "Name"},
-				new:      &Application{AppID: "AppID", Name: "NameChanged"},
-			},
-			res: res{
-				changesLen: 2,
-			},
-		},
-		{
-			name: "no changes",
-			args: args{
-				existing: &Application{AppID: "AppID", Name: "Name"},
-				new:      &Application{AppID: "AppID", Name: "Name"},
-			},
-			res: res{
-				changesLen: 1,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			changes := tt.args.existing.Changes(tt.args.new)
-			if len(changes) != tt.res.changesLen {
-				t.Errorf("got wrong changes len: expected: %v, actual: %v ", tt.res.changesLen, len(changes))
-			}
-		})
-	}
-}
-
-func TestOIDCConfigChanges(t *testing.T) {
-	type args struct {
-		existing *OIDCConfig
-		new      *OIDCConfig
-	}
-	type res struct {
-		changesLen int
-	}
-	tests := []struct {
-		name string
-		args args
-		res  res
-	}{
-		{
-			name: "all possible values change",
-			args: args{
-				existing: &OIDCConfig{
-					AppID:                  "AppID",
-					RedirectUris:           []string{"RedirectUris"},
-					ResponseTypes:          []int32{1},
-					GrantTypes:             []int32{1},
-					ApplicationType:        1,
-					AuthMethodType:         1,
-					PostLogoutRedirectUris: []string{"PostLogoutRedirectUris"},
-				},
-				new: &OIDCConfig{
-					AppID:                  "AppID",
-					RedirectUris:           []string{"RedirectUrisChanged"},
-					ResponseTypes:          []int32{2},
-					GrantTypes:             []int32{2},
-					ApplicationType:        2,
-					AuthMethodType:         2,
-					PostLogoutRedirectUris: []string{"PostLogoutRedirectUrisChanged"},
-				},
-			},
-			res: res{
-				changesLen: 7,
-			},
-		},
-		{
-			name: "no changes",
-			args: args{
-				existing: &OIDCConfig{
-					AppID:                  "AppID",
-					RedirectUris:           []string{"RedirectUris"},
-					ResponseTypes:          []int32{1},
-					GrantTypes:             []int32{1},
-					ApplicationType:        1,
-					AuthMethodType:         1,
-					PostLogoutRedirectUris: []string{"PostLogoutRedirectUris"},
-				},
-				new: &OIDCConfig{
-					AppID:                  "AppID",
-					RedirectUris:           []string{"RedirectUris"},
-					ResponseTypes:          []int32{1},
-					GrantTypes:             []int32{1},
-					ApplicationType:        1,
-					AuthMethodType:         1,
-					PostLogoutRedirectUris: []string{"PostLogoutRedirectUris"},
-				},
-			},
-			res: res{
-				changesLen: 1,
-			},
-		},
-		{
-			name: "change not changeable attributes",
-			args: args{
-				existing: &OIDCConfig{
-					AppID:    "AppID",
-					ClientID: "ClientID",
-				},
-				new: &OIDCConfig{
-					AppID:    "AppIDChange",
-					ClientID: "ClientIDChange",
-				},
-			},
-			res: res{
-				changesLen: 1,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			changes := tt.args.existing.Changes(tt.args.new)
-			if len(changes) != tt.res.changesLen {
-				t.Errorf("got wrong changes len: expected: %v, actual: %v ", tt.res.changesLen, len(changes))
-			}
-		})
-	}
-}
-
-func TestProjectFromEvents(t *testing.T) {
-	type args struct {
-		event   []*es_models.Event
-		project *Project
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "project from events, ok",
-			args: args{
-				event: []*es_models.Event{
-					&es_models.Event{AggregateID: "AggregateID", Sequence: 1, Type: model.ProjectAdded},
-				},
-				project: &Project{Name: "ProjectName"},
-			},
-			result: &Project{ObjectRoot: es_models.ObjectRoot{AggregateID: "AggregateID"}, State: int32(model.PROJECTSTATE_ACTIVE), Name: "ProjectName"},
-		},
-		{
-			name: "project from events, nil project",
-			args: args{
-				event: []*es_models.Event{
-					&es_models.Event{AggregateID: "AggregateID", Sequence: 1, Type: model.ProjectAdded},
-				},
-				project: nil,
-			},
-			result: &Project{ObjectRoot: es_models.ObjectRoot{AggregateID: "AggregateID"}, State: int32(model.PROJECTSTATE_ACTIVE)},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.project != nil {
-				data, _ := json.Marshal(tt.args.project)
-				tt.args.event[0].Data = data
-			}
-			result, _ := ProjectFromEvents(tt.args.project, tt.args.event...)
-			if result.Name != tt.result.Name {
-				t.Errorf("got wrong result name: expected: %v, actual: %v ", tt.result.Name, result.Name)
-			}
-		})
-	}
-}
-
-func TestAppendEvent(t *testing.T) {
-	type args struct {
-		event   *es_models.Event
-		project *Project
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append added event",
-			args: args{
-				event:   &es_models.Event{AggregateID: "AggregateID", Sequence: 1, Type: model.ProjectAdded},
-				project: &Project{Name: "ProjectName"},
-			},
-			result: &Project{ObjectRoot: es_models.ObjectRoot{AggregateID: "AggregateID"}, State: int32(model.PROJECTSTATE_ACTIVE), Name: "ProjectName"},
-		},
-		{
-			name: "append change event",
-			args: args{
-				event:   &es_models.Event{AggregateID: "AggregateID", Sequence: 1, Type: model.ProjectChanged},
-				project: &Project{Name: "ProjectName"},
-			},
-			result: &Project{ObjectRoot: es_models.ObjectRoot{AggregateID: "AggregateID"}, State: int32(model.PROJECTSTATE_ACTIVE), Name: "ProjectName"},
-		},
-		{
-			name: "append deactivate event",
-			args: args{
-				event: &es_models.Event{AggregateID: "AggregateID", Sequence: 1, Type: model.ProjectDeactivated},
-			},
-			result: &Project{ObjectRoot: es_models.ObjectRoot{AggregateID: "AggregateID"}, State: int32(model.PROJECTSTATE_INACTIVE)},
-		},
-		{
-			name: "append reactivate event",
-			args: args{
-				event: &es_models.Event{AggregateID: "AggregateID", Sequence: 1, Type: model.ProjectReactivated},
-			},
-			result: &Project{ObjectRoot: es_models.ObjectRoot{AggregateID: "AggregateID"}, State: int32(model.PROJECTSTATE_ACTIVE)},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.project != nil {
-				data, _ := json.Marshal(tt.args.project)
-				tt.args.event.Data = data
-			}
-			result := new(Project)
-			result.AppendEvent(tt.args.event)
-			if result.State != tt.result.State {
-				t.Errorf("got wrong result state: expected: %v, actual: %v ", tt.result.State, result.State)
-			}
-			if result.Name != tt.result.Name {
-				t.Errorf("got wrong result name: expected: %v, actual: %v ", tt.result.Name, result.Name)
-			}
-			if result.ObjectRoot.AggregateID != tt.result.ObjectRoot.AggregateID {
-				t.Errorf("got wrong result id: expected: %v, actual: %v ", tt.result.ObjectRoot.AggregateID, result.ObjectRoot.AggregateID)
-			}
-		})
-	}
-}
-
-func TestAppendDeactivatedEvent(t *testing.T) {
-	type args struct {
-		project *Project
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append reactivate event",
-			args: args{
-				project: &Project{},
-			},
-			result: &Project{State: int32(model.PROJECTSTATE_INACTIVE)},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tt.args.project.appendDeactivatedEvent()
-			if tt.args.project.State != tt.result.State {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result, tt.args.project)
-			}
-		})
-	}
-}
-
-func TestAppendReactivatedEvent(t *testing.T) {
-	type args struct {
-		project *Project
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append reactivate event",
-			args: args{
-				project: &Project{},
-			},
-			result: &Project{State: int32(model.PROJECTSTATE_ACTIVE)},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tt.args.project.appendReactivatedEvent()
-			if tt.args.project.State != tt.result.State {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result, tt.args.project)
-			}
-		})
-	}
-}
-
-func TestAppendAddMemberEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		member  *ProjectMember
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append add member event",
-			args: args{
-				project: &Project{},
-				member:  &ProjectMember{UserID: "UserID", Roles: []string{"Role"}},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Members: []*ProjectMember{&ProjectMember{UserID: "UserID", Roles: []string{"Role"}}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.member != nil {
-				data, _ := json.Marshal(tt.args.member)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendAddMemberEvent(tt.args.event)
-			if len(tt.args.project.Members) != 1 {
-				t.Errorf("got wrong result should have one member actual: %v ", len(tt.args.project.Members))
-			}
-			if tt.args.project.Members[0] == tt.result.Members[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Members[0], tt.args.project.Members[0])
-			}
-		})
-	}
-}
-
-func TestAppendChangeMemberEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		member  *ProjectMember
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append change member event",
-			args: args{
-				project: &Project{Members: []*ProjectMember{&ProjectMember{UserID: "UserID", Roles: []string{"Role"}}}},
-				member:  &ProjectMember{UserID: "UserID", Roles: []string{"ChangedRole"}},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Members: []*ProjectMember{&ProjectMember{UserID: "UserID", Roles: []string{"ChangedRole"}}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.member != nil {
-				data, _ := json.Marshal(tt.args.member)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendChangeMemberEvent(tt.args.event)
-			if len(tt.args.project.Members) != 1 {
-				t.Errorf("got wrong result should have one member actual: %v ", len(tt.args.project.Members))
-			}
-			if tt.args.project.Members[0] == tt.result.Members[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Members[0], tt.args.project.Members[0])
-			}
-		})
-	}
-}
-
-func TestAppendRemoveMemberEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		member  *ProjectMember
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append remove member event",
-			args: args{
-				project: &Project{Members: []*ProjectMember{&ProjectMember{UserID: "UserID", Roles: []string{"Role"}}}},
-				member:  &ProjectMember{UserID: "UserID"},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Members: []*ProjectMember{}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.member != nil {
-				data, _ := json.Marshal(tt.args.member)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendRemoveMemberEvent(tt.args.event)
-			if len(tt.args.project.Members) != 0 {
-				t.Errorf("got wrong result should have no member actual: %v ", len(tt.args.project.Members))
-			}
-		})
-	}
-}
-
-func TestAppendAddRoleEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		role    *ProjectRole
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append add role event",
-			args: args{
-				project: &Project{},
-				role:    &ProjectRole{Key: "Key", DisplayName: "DisplayName", Group: "Group"},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Roles: []*ProjectRole{&ProjectRole{Key: "Key", DisplayName: "DisplayName", Group: "Group"}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.role != nil {
-				data, _ := json.Marshal(tt.args.role)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendAddRoleEvent(tt.args.event)
-			if len(tt.args.project.Roles) != 1 {
-				t.Errorf("got wrong result should have one role actual: %v ", len(tt.args.project.Roles))
-			}
-			if tt.args.project.Roles[0] == tt.result.Roles[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Roles[0], tt.args.project.Roles[0])
-			}
-		})
-	}
-}
-
-func TestAppendChangeRoleEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		role    *ProjectRole
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append change role event",
-			args: args{
-				project: &Project{Roles: []*ProjectRole{&ProjectRole{Key: "Key", DisplayName: "DisplayName", Group: "Group"}}},
-				role:    &ProjectRole{Key: "Key", DisplayName: "DisplayNameChanged", Group: "Group"},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Roles: []*ProjectRole{&ProjectRole{Key: "Key", DisplayName: "DisplayNameChanged", Group: "Group"}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.role != nil {
-				data, _ := json.Marshal(tt.args.role)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendChangeRoleEvent(tt.args.event)
-			if len(tt.args.project.Roles) != 1 {
-				t.Errorf("got wrong result should have one role actual: %v ", len(tt.args.project.Roles))
-			}
-			if tt.args.project.Roles[0] == tt.result.Roles[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Roles[0], tt.args.project.Roles[0])
-			}
-		})
-	}
-}
-
-func TestAppendRemoveRoleEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		role    *ProjectRole
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append remove role event",
-			args: args{
-				project: &Project{Roles: []*ProjectRole{&ProjectRole{Key: "Key", DisplayName: "DisplayName", Group: "Group"}}},
-				role:    &ProjectRole{Key: "Key"},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Roles: []*ProjectRole{}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.role != nil {
-				data, _ := json.Marshal(tt.args.role)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendRemoveRoleEvent(tt.args.event)
-			if len(tt.args.project.Roles) != 0 {
-				t.Errorf("got wrong result should have no role actual: %v ", len(tt.args.project.Roles))
-			}
-		})
-	}
-}
-
-func TestAppendAddAppEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		app     *Application
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append add application event",
-			args: args{
-				project: &Project{},
-				app:     &Application{Name: "Application"},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Applications: []*Application{&Application{Name: "Application"}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.app != nil {
-				data, _ := json.Marshal(tt.args.app)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendAddAppEvent(tt.args.event)
-			if len(tt.args.project.Applications) != 1 {
-				t.Errorf("got wrong result should have one app actual: %v ", len(tt.args.project.Applications))
-			}
-			if tt.args.project.Applications[0] == tt.result.Applications[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Applications[0], tt.args.project.Applications[0])
-			}
-		})
-	}
-}
-
-func TestAppendChangeAppEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		app     *Application
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append change application event",
-			args: args{
-				project: &Project{Applications: []*Application{&Application{Name: "Application"}}},
-				app:     &Application{Name: "Application Change"},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Applications: []*Application{&Application{Name: "Application Change"}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.app != nil {
-				data, _ := json.Marshal(tt.args.app)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendChangeAppEvent(tt.args.event)
-			if len(tt.args.project.Applications) != 1 {
-				t.Errorf("got wrong result should have one app actual: %v ", len(tt.args.project.Applications))
-			}
-			if tt.args.project.Applications[0] == tt.result.Applications[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Applications[0], tt.args.project.Applications[0])
-			}
-		})
-	}
-}
-
-func TestAppendRemoveAppEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		app     *Application
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append remove application event",
-			args: args{
-				project: &Project{Applications: []*Application{&Application{AppID: "AppID", Name: "Application"}}},
-				app:     &Application{AppID: "AppID", Name: "Application"},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Applications: []*Application{}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.app != nil {
-				data, _ := json.Marshal(tt.args.app)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendRemoveAppEvent(tt.args.event)
-			if len(tt.args.project.Applications) != 0 {
-				t.Errorf("got wrong result should have no apps actual: %v ", len(tt.args.project.Applications))
-			}
-		})
-	}
-}
-
-func TestAppendAppStateEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		app     *ApplicationID
-		event   *es_models.Event
-		state   model.AppState
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append deactivate application event",
-			args: args{
-				project: &Project{Applications: []*Application{&Application{AppID: "AppID", Name: "Application", State: int32(model.APPSTATE_ACTIVE)}}},
-				app:     &ApplicationID{AppID: "AppID"},
-				event:   &es_models.Event{},
-				state:   model.APPSTATE_INACTIVE,
-			},
-			result: &Project{Applications: []*Application{&Application{AppID: "AppID", Name: "Application", State: int32(model.APPSTATE_INACTIVE)}}},
-		},
-		{
-			name: "append reactivate application event",
-			args: args{
-				project: &Project{Applications: []*Application{&Application{AppID: "AppID", Name: "Application", State: int32(model.APPSTATE_INACTIVE)}}},
-				app:     &ApplicationID{AppID: "AppID"},
-				event:   &es_models.Event{},
-				state:   model.APPSTATE_ACTIVE,
-			},
-			result: &Project{Applications: []*Application{&Application{AppID: "AppID", Name: "Application", State: int32(model.APPSTATE_ACTIVE)}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.app != nil {
-				data, _ := json.Marshal(tt.args.app)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendAppStateEvent(tt.args.event, tt.args.state)
-			if len(tt.args.project.Applications) != 1 {
-				t.Errorf("got wrong result should have one app actual: %v ", len(tt.args.project.Applications))
-			}
-			if tt.args.project.Applications[0] == tt.result.Applications[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Applications[0], tt.args.project.Applications[0])
-			}
-		})
-	}
-}
-
-func TestAppendAddOIDCConfigEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		config  *OIDCConfig
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append add application event",
-			args: args{
-				project: &Project{Applications: []*Application{&Application{AppID: "AppID"}}},
-				config:  &OIDCConfig{AppID: "AppID", ClientID: "ClientID"},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Applications: []*Application{&Application{AppID: "AppID", OIDCConfig: &OIDCConfig{AppID: "AppID", ClientID: "ClientID"}}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.config != nil {
-				data, _ := json.Marshal(tt.args.config)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendAddOIDCConfigEvent(tt.args.event)
-			if len(tt.args.project.Applications) != 1 {
-				t.Errorf("got wrong result should have one app actual: %v ", len(tt.args.project.Applications))
-			}
-			if tt.args.project.Applications[0].OIDCConfig == nil {
-				t.Errorf("got wrong result should have oidc config actual: %v ", tt.args.project.Applications[0].OIDCConfig)
-			}
-			if tt.args.project.Applications[0] == tt.result.Applications[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Applications[0], tt.args.project.Applications[0])
-			}
-		})
-	}
-}
-
-func TestAppendChangeOIDCConfigEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		config  *OIDCConfig
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append change application event",
-			args: args{
-				project: &Project{Applications: []*Application{&Application{AppID: "AppID", OIDCConfig: &OIDCConfig{AppID: "AppID", ClientID: "ClientID"}}}},
-				config:  &OIDCConfig{AppID: "AppID", ClientID: "ClientID Changed"},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Applications: []*Application{&Application{AppID: "AppID", OIDCConfig: &OIDCConfig{AppID: "AppID", ClientID: "ClientID Changed"}}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.config != nil {
-				data, _ := json.Marshal(tt.args.config)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendChangeOIDCConfigEvent(tt.args.event)
-			if len(tt.args.project.Applications) != 1 {
-				t.Errorf("got wrong result should have one app actual: %v ", len(tt.args.project.Applications))
-			}
-			if tt.args.project.Applications[0].OIDCConfig == nil {
-				t.Errorf("got wrong result should have oidc config actual: %v ", tt.args.project.Applications[0].OIDCConfig)
-			}
-			if tt.args.project.Applications[0] == tt.result.Applications[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Applications[0], tt.args.project.Applications[0])
-			}
-		})
-	}
-}
-
-func TestAppendAddApplicationEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		role    *ProjectGrant
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append add grant event",
-			args: args{
-				project: &Project{},
-				role:    &ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"Key"}},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Grants: []*ProjectGrant{&ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"Key"}}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.role != nil {
-				data, _ := json.Marshal(tt.args.role)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendAddGrantEvent(tt.args.event)
-			if len(tt.args.project.Grants) != 1 {
-				t.Errorf("got wrong result should have one grant actual: %v ", len(tt.args.project.Grants))
-			}
-			if tt.args.project.Grants[0] == tt.result.Grants[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Grants[0], tt.args.project.Grants[0])
-			}
-		})
-	}
-}
-
-func TestAppendChangeGrantEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		grant   *ProjectGrant
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append change grant event",
-			args: args{
-				project: &Project{Grants: []*ProjectGrant{&ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"Key"}}}},
-				grant:   &ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"KeyChanged"}},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Grants: []*ProjectGrant{&ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"KeyChanged"}}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.grant != nil {
-				data, _ := json.Marshal(tt.args.grant)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendChangeGrantEvent(tt.args.event)
-			if len(tt.args.project.Grants) != 1 {
-				t.Errorf("got wrong result should have one grant actual: %v ", len(tt.args.project.Grants))
-			}
-			if tt.args.project.Grants[0] == tt.result.Grants[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Grants[0], tt.args.project.Grants[0])
-			}
-		})
-	}
-}
-
-func TestAppendRemoveGrantEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		grant   *ProjectGrant
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append remove role event",
-			args: args{
-				project: &Project{Grants: []*ProjectGrant{&ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"Key"}}}},
-				grant:   &ProjectGrant{GrantID: "GrantID"},
-				event:   &es_models.Event{},
-			},
-			result: &Project{Grants: []*ProjectGrant{}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.grant != nil {
-				data, _ := json.Marshal(tt.args.grant)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendRemoveGrantEvent(tt.args.event)
-			if len(tt.args.project.Grants) != 0 {
-				t.Errorf("got wrong result should have no grant actual: %v ", len(tt.args.project.Grants))
-			}
-		})
-	}
-}
-
-func TestAppendGrantStateEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		grant   *ProjectGrantID
-		event   *es_models.Event
-		state   model.ProjectGrantState
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append deactivate grant event",
-			args: args{
-				project: &Project{Grants: []*ProjectGrant{&ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"Key"}}}},
-				grant:   &ProjectGrantID{GrantID: "GrantID"},
-				event:   &es_models.Event{},
-				state:   model.PROJECTGRANTSTATE_INACTIVE,
-			},
-			result: &Project{Grants: []*ProjectGrant{&ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"Key"}, State: int32(model.PROJECTGRANTSTATE_INACTIVE)}}},
-		},
-		{
-			name: "append reactivate grant event",
-			args: args{
-				project: &Project{Grants: []*ProjectGrant{&ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"Key"}}}},
-				grant:   &ProjectGrantID{GrantID: "GrantID"},
-				event:   &es_models.Event{},
-				state:   model.PROJECTGRANTSTATE_ACTIVE,
-			},
-			result: &Project{Grants: []*ProjectGrant{&ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"Key"}, State: int32(model.PROJECTGRANTSTATE_ACTIVE)}}},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.grant != nil {
-				data, _ := json.Marshal(tt.args.grant)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendGrantStateEvent(tt.args.event, tt.args.state)
-			if len(tt.args.project.Grants) != 1 {
-				t.Errorf("got wrong result should have one grant actual: %v ", len(tt.args.project.Grants))
-			}
-			if tt.args.project.Grants[0] == tt.result.Grants[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Grants[0], tt.args.project.Grants[0])
-			}
-		})
-	}
-}
-
-func TestAppendAddGrantMemberEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		member  *ProjectGrantMember
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append add grant member",
-			args: args{
-				project: &Project{Grants: []*ProjectGrant{
-					&ProjectGrant{GrantID: "GrantID", GrantedOrgID: "OrgID", RoleKeys: []string{"Key"}}}},
-				member: &ProjectGrantMember{GrantID: "GrantID", UserID: "UserID", Roles: []string{"Role"}},
-				event:  &es_models.Event{},
-			},
-			result: &Project{
-				Grants: []*ProjectGrant{
-					&ProjectGrant{
-						GrantID:      "GrantID",
-						GrantedOrgID: "OrgID",
-						RoleKeys:     []string{"Key"},
-						Members:      []*ProjectGrantMember{&ProjectGrantMember{GrantID: "GrantID", UserID: "UserID", Roles: []string{"Role"}}}}},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.member != nil {
-				data, _ := json.Marshal(tt.args.member)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendAddGrantMemberEvent(tt.args.event)
-			if len(tt.args.project.Grants[0].Members) != 1 {
-				t.Errorf("got wrong result should have one grant actual: %v ", len(tt.args.project.Grants[0].Members))
-			}
-			if tt.args.project.Grants[0].Members[0] == tt.result.Grants[0].Members[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Grants[0].Members[0], tt.args.project.Grants[0].Members[0])
-			}
-		})
-	}
-}
-
-func TestAppendChangeGrantMemberEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		member  *ProjectGrantMember
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name   string
-		args   args
-		result *Project
-	}{
-		{
-			name: "append change grant member",
-			args: args{
-				project: &Project{
-					Grants: []*ProjectGrant{
-						&ProjectGrant{
-							GrantID:      "GrantID",
-							GrantedOrgID: "OrgID",
-							RoleKeys:     []string{"Key"},
-							Members:      []*ProjectGrantMember{&ProjectGrantMember{GrantID: "GrantID", UserID: "UserID", Roles: []string{"Role"}}}}},
-				},
-				member: &ProjectGrantMember{GrantID: "GrantID", UserID: "UserID", Roles: []string{"RoleChanged"}},
-				event:  &es_models.Event{},
-			},
-			result: &Project{
-				Grants: []*ProjectGrant{
-					&ProjectGrant{
-						GrantID:      "GrantID",
-						GrantedOrgID: "OrgID",
-						RoleKeys:     []string{"Key"},
-						Members:      []*ProjectGrantMember{&ProjectGrantMember{GrantID: "GrantID", UserID: "UserID", Roles: []string{"RoleChanged"}}}}},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.member != nil {
-				data, _ := json.Marshal(tt.args.member)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendChangeGrantMemberEvent(tt.args.event)
-			if len(tt.args.project.Grants[0].Members) != 1 {
-				t.Errorf("got wrong result should have one grant actual: %v ", len(tt.args.project.Grants[0].Members))
-			}
-			if tt.args.project.Grants[0].Members[0] == tt.result.Grants[0].Members[0] {
-				t.Errorf("got wrong result: expected: %v, actual: %v ", tt.result.Grants[0].Members[0], tt.args.project.Grants[0].Members[0])
-			}
-		})
-	}
-}
-
-func TestAppendRemoveGrantMemberEvent(t *testing.T) {
-	type args struct {
-		project *Project
-		member  *ProjectGrantMember
-		event   *es_models.Event
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "append remove grant member",
-			args: args{
-				project: &Project{
-					Grants: []*ProjectGrant{
-						&ProjectGrant{
-							GrantID:      "GrantID",
-							GrantedOrgID: "OrgID",
-							RoleKeys:     []string{"Key"},
-							Members:      []*ProjectGrantMember{&ProjectGrantMember{GrantID: "GrantID", UserID: "UserID", Roles: []string{"Role"}}}}},
-				},
-				member: &ProjectGrantMember{GrantID: "GrantID", UserID: "UserID", Roles: []string{"RoleChanged"}},
-				event:  &es_models.Event{},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.member != nil {
-				data, _ := json.Marshal(tt.args.member)
-				tt.args.event.Data = data
-			}
-			tt.args.project.appendRemoveGrantMemberEvent(tt.args.event)
-			if len(tt.args.project.Grants[0].Members) != 0 {
-				t.Errorf("got wrong result should have no members actual: %v ", len(tt.args.project.Grants[0].Members))
-			}
-		})
-	}
-}
