+++ internal/project/repository/eventsourcing/project.go
@@ -7,6 +7,7 @@ import (
 	"github.com/caos/zitadel/internal/eventstore/models"
 	es_models "github.com/caos/zitadel/internal/eventstore/models"
 	"github.com/caos/zitadel/internal/project/model"
+	model2 "github.com/caos/zitadel/internal/project/repository/eventsourcing/model"
 )
 
 func ProjectByIDQuery(id string, latestSequence uint64) (*es_models.SearchQuery, error) {
@@ -23,14 +24,14 @@ func ProjectQuery(latestSequence uint64) *es_models.SearchQuery {
 		LatestSequenceFilter(latestSequence)
 }
 
-func ProjectAggregate(ctx context.Context, aggCreator *es_models.AggregateCreator, project *Project) (*es_models.Aggregate, error) {
+func ProjectAggregate(ctx context.Context, aggCreator *es_models.AggregateCreator, project *model2.Project) (*es_models.Aggregate, error) {
 	if project == nil {
 		return nil, errors.ThrowPreconditionFailed(nil, "EVENT-doe93", "existing project should not be nil")
 	}
-	return aggCreator.NewAggregate(ctx, project.AggregateID, model.ProjectAggregate, projectVersion, project.Sequence)
+	return aggCreator.NewAggregate(ctx, project.AggregateID, model.ProjectAggregate, model2.ProjectVersion, project.Sequence)
 }
 
-func ProjectCreateAggregate(aggCreator *es_models.AggregateCreator, project *Project) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectCreateAggregate(aggCreator *es_models.AggregateCreator, project *model2.Project) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if project == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-kdie6", "project should not be nil")
@@ -45,7 +46,7 @@ func ProjectCreateAggregate(aggCreator *es_models.AggregateCreator, project *Pro
 	}
 }
 
-func ProjectUpdateAggregate(aggCreator *es_models.AggregateCreator, existing *Project, new *Project) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectUpdateAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, new *model2.Project) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if new == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-dhr74", "new project should not be nil")
@@ -59,15 +60,15 @@ func ProjectUpdateAggregate(aggCreator *es_models.AggregateCreator, existing *Pr
 	}
 }
 
-func ProjectDeactivateAggregate(aggCreator *es_models.AggregateCreator, project *Project) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectDeactivateAggregate(aggCreator *es_models.AggregateCreator, project *model2.Project) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return projectStateAggregate(aggCreator, project, model.ProjectDeactivated)
 }
 
-func ProjectReactivateAggregate(aggCreator *es_models.AggregateCreator, project *Project) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectReactivateAggregate(aggCreator *es_models.AggregateCreator, project *model2.Project) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return projectStateAggregate(aggCreator, project, model.ProjectReactivated)
 }
 
-func projectStateAggregate(aggCreator *es_models.AggregateCreator, project *Project, state models.EventType) func(ctx context.Context) (*es_models.Aggregate, error) {
+func projectStateAggregate(aggCreator *es_models.AggregateCreator, project *model2.Project, state models.EventType) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		agg, err := ProjectAggregate(ctx, aggCreator, project)
 		if err != nil {
@@ -77,7 +78,7 @@ func projectStateAggregate(aggCreator *es_models.AggregateCreator, project *Proj
 	}
 }
 
-func ProjectMemberAddedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, member *ProjectMember) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectMemberAddedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, member *model2.ProjectMember) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if member == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-ie34f", "member should not be nil")
@@ -90,7 +91,7 @@ func ProjectMemberAddedAggregate(aggCreator *es_models.AggregateCreator, existin
 	}
 }
 
-func ProjectMemberChangedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, member *ProjectMember) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectMemberChangedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, member *model2.ProjectMember) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if member == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-d34fs", "member should not be nil")
@@ -104,7 +105,7 @@ func ProjectMemberChangedAggregate(aggCreator *es_models.AggregateCreator, exist
 	}
 }
 
-func ProjectMemberRemovedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, member *ProjectMember) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectMemberRemovedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, member *model2.ProjectMember) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if member == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-dieu7", "member should not be nil")
@@ -117,7 +118,7 @@ func ProjectMemberRemovedAggregate(aggCreator *es_models.AggregateCreator, exist
 	}
 }
 
-func ProjectRoleAddedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, role *ProjectRole) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectRoleAddedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, role *model2.ProjectRole) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if role == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-sleo9", "role should not be nil")
@@ -130,7 +131,7 @@ func ProjectRoleAddedAggregate(aggCreator *es_models.AggregateCreator, existing 
 	}
 }
 
-func ProjectRoleChangedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, role *ProjectRole) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectRoleChangedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, role *model2.ProjectRole) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if role == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-oe8sf", "member should not be nil")
@@ -143,7 +144,7 @@ func ProjectRoleChangedAggregate(aggCreator *es_models.AggregateCreator, existin
 	}
 }
 
-func ProjectRoleRemovedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, role *ProjectRole) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectRoleRemovedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, role *model2.ProjectRole) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if role == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-d8eis", "member should not be nil")
@@ -156,7 +157,7 @@ func ProjectRoleRemovedAggregate(aggCreator *es_models.AggregateCreator, existin
 	}
 }
 
-func ApplicationAddedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, app *Application) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ApplicationAddedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, app *model2.Application) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if app == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-09du7", "app should not be nil")
@@ -173,7 +174,7 @@ func ApplicationAddedAggregate(aggCreator *es_models.AggregateCreator, existing 
 	}
 }
 
-func ApplicationChangedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, app *Application) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ApplicationChangedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, app *model2.Application) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if app == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-sleo9", "app should not be nil")
@@ -194,7 +195,7 @@ func ApplicationChangedAggregate(aggCreator *es_models.AggregateCreator, existin
 	}
 }
 
-func ApplicationRemovedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, app *Application) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ApplicationRemovedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, app *model2.Application) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if app == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-se23g", "app should not be nil")
@@ -203,13 +204,13 @@ func ApplicationRemovedAggregate(aggCreator *es_models.AggregateCreator, existin
 		if err != nil {
 			return nil, err
 		}
-		agg.AppendEvent(model.ApplicationRemoved, &ApplicationID{AppID: app.AppID})
+		agg.AppendEvent(model.ApplicationRemoved, &model2.ApplicationID{AppID: app.AppID})
 
 		return agg, nil
 	}
 }
 
-func ApplicationDeactivatedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, app *Application) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ApplicationDeactivatedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, app *model2.Application) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if app == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-slfi3", "app should not be nil")
@@ -218,13 +219,13 @@ func ApplicationDeactivatedAggregate(aggCreator *es_models.AggregateCreator, exi
 		if err != nil {
 			return nil, err
 		}
-		agg.AppendEvent(model.ApplicationDeactivated, &ApplicationID{AppID: app.AppID})
+		agg.AppendEvent(model.ApplicationDeactivated, &model2.ApplicationID{AppID: app.AppID})
 
 		return agg, nil
 	}
 }
 
-func ApplicationReactivatedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, app *Application) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ApplicationReactivatedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, app *model2.Application) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if app == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-slf32", "app should not be nil")
@@ -233,13 +234,13 @@ func ApplicationReactivatedAggregate(aggCreator *es_models.AggregateCreator, exi
 		if err != nil {
 			return nil, err
 		}
-		agg.AppendEvent(model.ApplicationReactivated, &ApplicationID{AppID: app.AppID})
+		agg.AppendEvent(model.ApplicationReactivated, &model2.ApplicationID{AppID: app.AppID})
 
 		return agg, nil
 	}
 }
 
-func OIDCConfigChangedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, config *OIDCConfig) func(ctx context.Context) (*es_models.Aggregate, error) {
+func OIDCConfigChangedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, config *model2.OIDCConfig) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if config == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-slf32", "config should not be nil")
@@ -262,7 +263,7 @@ func OIDCConfigChangedAggregate(aggCreator *es_models.AggregateCreator, existing
 	}
 }
 
-func OIDCConfigSecretChangedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, appID string, secret *crypto.CryptoValue) func(ctx context.Context) (*es_models.Aggregate, error) {
+func OIDCConfigSecretChangedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, appID string, secret *crypto.CryptoValue) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		agg, err := ProjectAggregate(ctx, aggCreator, existing)
 		if err != nil {
@@ -278,7 +279,7 @@ func OIDCConfigSecretChangedAggregate(aggCreator *es_models.AggregateCreator, ex
 	}
 }
 
-func ProjectGrantAddedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, grant *ProjectGrant) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectGrantAddedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, grant *model2.ProjectGrant) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if grant == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-kd89w", "grant should not be nil")
@@ -292,7 +293,7 @@ func ProjectGrantAddedAggregate(aggCreator *es_models.AggregateCreator, existing
 	}
 }
 
-func ProjectGrantChangedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, grant *ProjectGrant) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectGrantChangedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, grant *model2.ProjectGrant) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if grant == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-d9ie2", "grant should not be nil")
@@ -313,7 +314,7 @@ func ProjectGrantChangedAggregate(aggCreator *es_models.AggregateCreator, existi
 	}
 }
 
-func ProjectGrantRemovedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, grant *ProjectGrant) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectGrantRemovedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, grant *model2.ProjectGrant) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if grant == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-kci8d", "grant should not be nil")
@@ -322,13 +323,13 @@ func ProjectGrantRemovedAggregate(aggCreator *es_models.AggregateCreator, existi
 		if err != nil {
 			return nil, err
 		}
-		agg.AppendEvent(model.ProjectGrantRemoved, &ProjectGrantID{GrantID: grant.GrantID})
+		agg.AppendEvent(model.ProjectGrantRemoved, &model2.ProjectGrantID{GrantID: grant.GrantID})
 
 		return agg, nil
 	}
 }
 
-func ProjectGrantDeactivatedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, grant *ProjectGrant) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectGrantDeactivatedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, grant *model2.ProjectGrant) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if grant == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-id832", "grant should not be nil")
@@ -337,13 +338,13 @@ func ProjectGrantDeactivatedAggregate(aggCreator *es_models.AggregateCreator, ex
 		if err != nil {
 			return nil, err
 		}
-		agg.AppendEvent(model.ProjectGrantDeactivated, &ProjectGrantID{GrantID: grant.GrantID})
+		agg.AppendEvent(model.ProjectGrantDeactivated, &model2.ProjectGrantID{GrantID: grant.GrantID})
 
 		return agg, nil
 	}
 }
 
-func ProjectGrantReactivatedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, grant *ProjectGrant) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectGrantReactivatedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, grant *model2.ProjectGrant) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if grant == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-8diw2", "grant should not be nil")
@@ -352,13 +353,13 @@ func ProjectGrantReactivatedAggregate(aggCreator *es_models.AggregateCreator, ex
 		if err != nil {
 			return nil, err
 		}
-		agg.AppendEvent(model.ProjectGrantReactivated, &ProjectGrantID{GrantID: grant.GrantID})
+		agg.AppendEvent(model.ProjectGrantReactivated, &model2.ProjectGrantID{GrantID: grant.GrantID})
 
 		return agg, nil
 	}
 }
 
-func ProjectGrantMemberAddedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, member *ProjectGrantMember) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectGrantMemberAddedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, member *model2.ProjectGrantMember) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if member == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-4ufh6", "grant should not be nil")
@@ -372,7 +373,7 @@ func ProjectGrantMemberAddedAggregate(aggCreator *es_models.AggregateCreator, ex
 	}
 }
 
-func ProjectGrantMemberChangedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, member *ProjectGrantMember) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectGrantMemberChangedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, member *model2.ProjectGrantMember) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if member == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-d8i4h", "member should not be nil")
@@ -391,7 +392,7 @@ func ProjectGrantMemberChangedAggregate(aggCreator *es_models.AggregateCreator, 
 	}
 }
 
-func ProjectGrantMemberRemovedAggregate(aggCreator *es_models.AggregateCreator, existing *Project, member *ProjectGrantMember) func(ctx context.Context) (*es_models.Aggregate, error) {
+func ProjectGrantMemberRemovedAggregate(aggCreator *es_models.AggregateCreator, existing *model2.Project, member *model2.ProjectGrantMember) func(ctx context.Context) (*es_models.Aggregate, error) {
 	return func(ctx context.Context) (*es_models.Aggregate, error) {
 		if member == nil {
 			return nil, errors.ThrowPreconditionFailed(nil, "EVENT-slp0r", "member should not be nil")
